+++ HISTORY.md
@@ -8,6 +8,7 @@
 * objects: made color swatch background white for (semi-) transparent colors
 * threads: hyper-dyadicized "attribute of color" reporter
 * objects: visualize fully transparent colors as light gray checkerboard in the color swatch morph
+* objects, threads: added "new color h:s:v:" reporter primitive in the pen category
 
 ### 2025-04-02
 * blocks, objects: added "color" selector to pen-attributes reporter